feat: add room database dep and architecture
