+++ app/build.gradle
@@ -1,6 +1,7 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
 
 android {
     compileSdkVersion 29
@@ -42,5 +43,12 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-livedata-ktx:$lifecycle_version"
     implementation "androidx.activity:activity-ktx:1.2.0"
-
+    implementation 'com.github.kittinunf.fuel:fuel:2.2.1' //Core package
+    implementation 'com.github.kittinunf.fuel:fuel-android:2.2.1' //Android
+    implementation 'com.github.kittinunf.fuel:fuel-gson:2.2.1' //Fuel Gson
+    implementation 'com.google.code.gson:gson:2.8.5' //Gson
+    def room_version = "2.2.6"
+    implementation "androidx.room:room-runtime:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
+    implementation "androidx.room:room-ktx:$room_version"
 }
