+++ app/src/main/java/com/codingfactoryprojet/scanneropenfoodfact/database/AppDatabase.kt
@@ -1,0 +1,4 @@
+package com.codingfactoryprojet.scanneropenfoodfact.database
+
+class AppDatabase {
+}
